+++ application/controllers/Departamento.php
@@ -77,7 +77,7 @@ class Departamento extends CI_Controller
             $this->base->update('departamento', array('id' => $id), $data);
 
             //display success message
-            $this->session->set_flashdata('msg', '<div class="alert alert-success text-center">departamento actualizado exitosamente</div>');
+            $this->session->set_flashdata('msg', '<div class="alert alert-success fade in text-center"><a href="#" class="close" data-dismiss="alert">&times;</a>Departamento modificado exitosamente!</div>');
 
             redirect('departamento');
         }
@@ -104,7 +104,7 @@ class Departamento extends CI_Controller
             $this->base->add('departamento', $data);
 
             //display success message
-            $this->session->set_flashdata('msg', '<div class="alert alert-success text-center">Departamento registrado exitosamente</div>');
+            $this->session->set_flashdata('msg', '<div class="alert alert-success fade in text-center"><a href="#" class="close" data-dismiss="alert">&times;</a>Departamento registrado exitosamente!</div>');
 
             redirect('departamento');
         }
@@ -112,7 +112,13 @@ class Departamento extends CI_Controller
 
     public function delete($id){
 
-        $this->base->delete('departamento', array('id' => $id));
+        if ( $this->base->get('empleado', array('departamento_id' => $id)) ) {
+            //Si existen relaciones no eliminar
+            $this->session->set_flashdata('msg', '<div class="alert alert-danger fade in text-center"><a href="#" class="close" data-dismiss="alert">&times;</a>El departamento a eliminar tiene empleados registrados!</div>');
+        } else {
+            $this->base->delete('departamento', array('id' => $id));
+        }
+
         redirect('departamento');
     }
 }
